pyproject.toml
[tool.poetry]
name = "lms-drf-project"
version = "0.1.0"
description = "Learning Management System (LMS) built with Django REST Framework"
authors = ["GeeMoowy <kovylek.ul@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
pillow = "^11.2.1"
django-filter = "^25.1"
django = "^5.2.1"
djangorestframework-simplejwt = "^5.5.0"
coverage = "^7.8.2"
drf-yasg = "^1.21.10"
stripe = "^12.2.0"
flake8 = "^7.2.0"
celery = "^5.5.3"
eventlet = "^0.40.0"
redis = "^6.2.0"
django-celery-beat = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
